% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRB.r
\name{cvRB}
\alias{cvRB}
\title{Coefficient of Variation of the Redistributed-Beta Distribution}
\usage{
cvRB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical coefficient of variation of the RB distribution.
}
\examples{
cvRB(alpha = 2, beta = 5)

}
