% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRB.r
\name{kurtRB}
\alias{kurtRB}
\title{Kurtosis of the Redistributed-Beta Distribution}
\usage{
kurtRB(alpha, beta, excess = TRUE)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}

\item{excess}{Logical. If \code{TRUE}, the excess kurtosis is returned;
otherwise, the ordinary kurtosis is returned.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical kurtosis or excess kurtosis of the RB distribution.
}
\examples{
kurtRB(alpha = 2, beta = 5)
kurtRB(alpha = 2, beta = 5, excess = FALSE)

}
