% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRB.r
\name{meanRB}
\alias{meanRB}
\title{Mean of the Redistributed-Beta Distribution}
\usage{
meanRB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical mean of the RB distribution.
}
\examples{
meanRB(alpha = 2, beta = 5)

}
