% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRB.r
\name{skewRB}
\alias{skewRB}
\title{Skewness of the Redistributed-Beta Distribution}
\usage{
skewRB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical skewness coefficient of the RB distribution.
}
\examples{
skewRB(alpha = 2, beta = 5)

}
