% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRB.r
\name{varRB}
\alias{varRB}
\title{Variance of the Redistributed-Beta Distribution}
\usage{
varRB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical variance of the RB distribution.
}
\examples{
varRB(alpha = 2, beta = 5)

}
